Pulled code out of USBDeviceManager.cpp (was getting too big) * Moved USBBulkTransferBundle class into Device/USB/USBBulkTransferBundle.h/.cpp * Moved USBDeviceInfo.h into Device/USB/USBDeviceInfo.h * Moved USBRequest structs/enum into Device/USBDeviceRequest.h * Renamed USBRequestPayload_StartBulkTransfer -> USBRequestPayload_BulkTransfer * Renamed USBRequestPayload_StopBulkTransfer -> USBRequestPayload_CancelBulkTransfer
