+++ src/psmoveservice/CMakeLists.txt
@@ -65,6 +65,12 @@ file(GLOB PSMOVESERVICE_DEVICE_MGR_SRC
 )
 source_group("Device\\Manager" FILES ${PSMOVESERVICE_DEVICE_MGR_SRC})
 
+file(GLOB PSMOVESERVICE_DEVICE_USB_SRC
+    "${CMAKE_CURRENT_LIST_DIR}/Device/USB/*.cpp"
+    "${CMAKE_CURRENT_LIST_DIR}/Device/USB/*.h"    
+)
+source_group("Device\\USB" FILES ${PSMOVESERVICE_DEVICE_USB_SRC})
+
 file(GLOB PSMOVESERVICE_DEVICE_VIEW_SRC
     "${CMAKE_CURRENT_LIST_DIR}/Device/View/*.cpp"
     "${CMAKE_CURRENT_LIST_DIR}/Device/View/*.h"    
@@ -97,6 +103,7 @@ set(PSMOVESERVICE_SRC
     ${PSMOVESERVICE_DEVICE_ENUM_SRC}
     ${PSMOVESERVICE_DEVICE_INT_SRC}
     ${PSMOVESERVICE_DEVICE_MGR_SRC}
+    ${PSMOVESERVICE_DEVICE_USB_SRC}
     ${PSMOVESERVICE_DEVICE_VIEW_SRC}
     ${PSMOVESERVICE_HMD_SRC}
     ${PSMOVESERVICE_FILTER_SRC}
@@ -108,6 +115,7 @@ list(APPEND PSMOVE_SERVICE_INCL_DIRS
     ${CMAKE_CURRENT_LIST_DIR}/Device/Enumerator
     ${CMAKE_CURRENT_LIST_DIR}/Device/Interface
     ${CMAKE_CURRENT_LIST_DIR}/Device/Manager
+    ${CMAKE_CURRENT_LIST_DIR}/Device/USB
     ${CMAKE_CURRENT_LIST_DIR}/Device/View
     ${CMAKE_CURRENT_LIST_DIR}/Filter
     ${CMAKE_CURRENT_LIST_DIR}/OculusHMD
